% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bodymap}
\alias{bodymap}
\title{A Generic Body Map}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 14 rows and 4 columns.
}
\usage{
bodymap
}
\description{
A simple features data object for plotting body-related frequencies or
intensities.
}
\details{
A simple features data frame with 10 rows and 4 columns:
\describe{
  \item{id}{Unique body part ID}
  \item{NAME}{Body part Name}
  \item{SUBREGION}{Body part subregion (currently just used for the back)}
  \item{geometry}{Simploe features geometry column}
}
}
\keyword{datasets}
